# api/src/__generated__/openapi.yaml
openapi: 3.0.0
paths:
...
  /owners/{ownername}:
    get:
      description: User情報を取得します。
      operationId: OwnersController_findOne
      parameters:
        - name: ownername
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: 取得成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    description: ステータスコード
                    type: number
                  message:
                    description: メッセージ
                    type: string
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      username:
                        type: string
                      email:
                        type: string
                        format: email
                      fullname:
                        type: string
                      avatarUrl:
                        oneOf:
                          - type: string
                            format: uri
                          - type: string
                            enum:
                              - ''
                    required:
                      - id
                      - email
                      - avatarUrl
                required:
                  - statusCode
                  - message
                  - data
      security: &ref_0
        - supabase-auth: []
      summary: OwnersController_findOne / User情報の取得
      tags: &ref_1
        - owners
    put:
      operationId: OwnersController_update
      parameters:
        - name: ownername
          required: true
          in: path
          description: Owner名
          schema:
            example: user1
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ownerName:
                  type: string
                  minLength: 1
                username:
                  type: string
                  minLength: 1
                fullname:
                  type: string
                  minLength: 1
                email:
                  type: string
                  format: email
                avatarUrl:
                  oneOf:
                    - type: string
                      format: uri
                    - type: string
                      enum:
                        - ''
                readme:
                  type: string
              required:
                - ownerName
                - username
                - fullname
                - email
                - avatarUrl
                - readme
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    description: ステータスコード
                    type: number
                  message:
                    description: メッセージ
                    type: string
                required:
                  - statusCode
                  - message
      security: *ref_0
      summary: OwnersController_update / Owner情報の更新
      tags: *ref_1
  /owners/{ownername}/icon:
    get:
      operationId: OwnersController_getUserAvatar
      parameters:
        - name: ownername
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: 取得成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    description: ステータスコード
                    type: number
                  message:
                    description: メッセージ
                    type: string
                  data:
                    type: object
                    properties:
                      avatarUrl:
                        oneOf:
                          - type: string
                            format: uri
                          - type: string
                            enum:
                              - ''
                    required:
                      - avatarUrl
                required:
                  - statusCode
                  - message
                  - data
      security: *ref_0
      summary: OwnersController_getUserAvatar / UserのアバターURLを取得
      tags: *ref_1
    put:
      operationId: OwnersController_uploadUserIcon
      parameters:
        - name: ownername
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                file: {}
              required:
                - file
      responses:
        '200':
          description: アップロード成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    description: ステータスコード
                    type: number
                  message:
                    description: メッセージ
                    type: string
                required:
                  - statusCode
                  - message
      security: *ref_0
      summary: OwnersController_uploadUserIcon / アイコン画像をアップロード
      tags: *ref_1
  /owners/{ownername}/managepin:
    put:
      operationId: OwnersController_updatePin
      parameters:
        - name: ownername
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                projectId:
                  type: string
                  format: uuid
                pined:
                  type: boolean
              required:
                - projectId
                - pined
      responses:
        '200':
          description: ピン状態の更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    description: ステータスコード
                    type: number
                  message:
                    description: メッセージ
                    type: string
                required:
                  - statusCode
                  - message
      security: *ref_0
      summary: OwnersController_updatePin / Projectのピン状態を更新
      tags: *ref_1
    delete:
      operationId: OwnersController_leaveProject
      parameters:
        - name: ownername
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                projectId:
                  type: string
                  format: uuid
              required:
                - projectId
      responses:
        '200':
          description: 脱退成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    description: ステータスコード
                    type: number
                  message:
                    description: メッセージ
                    type: string
                required:
                  - statusCode
                  - message
      security: *ref_0
      summary: OwnersController_leaveProject / Projectから脱退
      tags: *ref_1
  /owners/{ownername}/orgdisplay:
    put:
      operationId: OwnersController_updateOrgDisplay
      parameters:
        - name: ownername
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                orgId:
                  type: string
                  format: uuid
                display:
                  type: boolean
              required:
                - orgId
                - display
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    description: ステータスコード
                    type: number
                  message:
                    description: メッセージ
                    type: string
                required:
                  - statusCode
                  - message
      security: *ref_0
      summary: OwnersController_updateOrgDisplay / 組織の表示・非表示の更新
      tags: *ref_1
    delete:
      operationId: OwnersController_leaveOrg
      parameters:
        - name: ownername
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                orgId:
                  type: string
                  format: uuid
              required:
                - orgId
      responses:
        '200':
          description: 脱退成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    description: ステータスコード
                    type: number
                  message:
                    description: メッセージ
                    type: string
                required:
                  - statusCode
                  - message
      security: *ref_0
      summary: OwnersController_leaveOrg / 組織から脱退
      tags: *ref_1
...
info:
  title: andTopic API
  description: andTopic API description
  version: '1.0'
  contact: {}
...
servers: []
...